Add ability to select and modify aircraft config values (no persistence)
This is a rough implementation of value editing in menus, with the responsibility
of value manipulation left to the view to handle itself (rather than implenting
a general increment/decrement for values, which may come later).

This currently selects config value based on the cursor index of different
settings. This saves 30 bytes here compared to comparing pointers to progmem
strings, but the readability hit isn't ideal.

Other notes:
- This saves 600 bytes of program space by using drawBox instead of drawRBox.
  I'd missed the former in the docs originally. The latter includes a whole
  lot of code for drawing rounded corners, which we don't need here.

+++ firmware/views/menu_shared.c
@@ -8,7 +8,7 @@
 
 static menu_item _shared_items[SHARED_MENU_SIZE];
 
-menu_item* populate_shared_menu(const menu_item* items, size_t count)
+menu_item* populate_shared_menu(const menu_item* items, uint8_t count)
 {
 	assert(count <= SHARED_MENU_SIZE);
